Add support for local check
Make the CLI command support an additional mode, which does not
check a PR, but instead checks the local Git state.

In this mode, not all checks can be executed, as some require
a pull request. There are a couple of checks, however,
namely those that work on the current branch and commits,
that can run without the need of a PR.