Fix number of commits checked locally
We only want the commits of the current branch from the parent
to the tip of the branch, e.g. master...my-feature-branch.

Before it would check all commits of the tree, starting from the
current branch and going all the way back to the initial commit of the
repository. Now it will only check the commits of the current branch.