Fix exception on merge commits
When running on a merge commit, there is no PR URL anymore.
Therefore, the command uses an empty string as a URL.
The way the command was, the value of the `--pr-url` argument
became `--config-file`, which was neither an empty string
nor a valid value.

Because it wasn't blank, the PRCheck run. But the URL was not
valid, so an exception occurred when trying to parse it.