Avoid running Totem on master and devel
Previously Totem ran whenever there was a new commit, including when
a PR was merged into the base branch. Because in that case there was
no pull request info in CircleCI's $PULL_REQUEST, Totem used a mode
that checked all commits of that branch. Because the branch would
probably include merge commits, the commit messages were too long
and Totem failed (e.g. "Merge pull request #108 from <branch-name>").

Now the CI script does not run Totem at all if the current branch
is either `devel` or `master`, so we won't be having problems
with Totem failing after merging a PR on those branches.

Note that this issue will still occur when merging a PR on other
branches. As this happens much more rarely in our current workflow,
this is not considered to be an issue.